cli/pyproject.toml
[project]
name = "campfire-cli"
version = "0.1.0"
description = "CLI tool for curating weekly link collections in Campfire"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
    "pillow>=12.0.0",
    "playwright>=1.58.0",
]

[project.scripts]
campfire-cli = "campfire_cli.main:app"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
